Allow 15char IP address
Last character in a char array string is always \0, so when mqtt_server array is 15 long it will discard the last digit in a 15 char ip address like 192.192.192.192

The mqtt socket time out is 15s by default, so to avoid an incredibly un responsive web interface when not connected to MQTT, the connection attempts are limited to once every 30s.

I considered setting the MQTT socket timeout to something like 5s, but I am not sure if that will create issues for some users.

Maybe we could set the MQTT socket timeout to 5s, and then add an input option so people can increase it in case 5s is not enough on their setup.